Updated Parcel class - removed available classmethod - removed oneline_address str property - removed coordinates tuple property - removed address dict property - added inventory classmethod - added address str property - added is_available bool property - changed: __str__ prints parcel_id and address
The Parcel class previously called a geocoder API in the coordinates
property. This increases the latency for determining the coordinates
for group of parcels substantially. Geocoding also violates seperation
of concerns for the Parcel class, meaning parcel is too entangled with
the gecoding API. Problems arise when geocoding changes and then
parcel has to change to track those changes. This cascading set of
changes is often referred to as "software fragility".